chore: Convert sheets for "Now Playing" screen to use refs
