+++ mobile/src/screens/NowPlaying/Sheets.tsx
@@ -1,26 +1,104 @@
-import { useEffect } from "react";
 import { useTranslation } from "react-i18next";
+import type { ActionSheetRef } from "react-native-actions-sheet";
+import TrackPlayer from "react-native-track-player";
 
 import { getTrackCover } from "~/db/utils";
 
 import { Remove } from "~/icons/Remove";
+import { SlowMotionVideo } from "~/icons/SlowMotionVideo";
+import { VolumeUp } from "~/icons/VolumeUp";
 import { Queue, useMusicStore } from "~/modules/media/services/Music";
-import { useUpcomingStore } from "./store";
+import { sessionStore, useSessionStore } from "~/services/SessionStore";
+import { useUpcomingStore } from "./helpers/UpcomingStore";
 
 import { Colors } from "~/constants/Styles";
 import { cn } from "~/lib/style";
 import { FlashList, SheetsFlashList } from "~/components/Defaults";
 import { IconButton } from "~/components/Form/Button";
+import { NSlider } from "~/components/Form/Slider";
 import { Sheet } from "~/components/Sheet";
 import { Swipeable } from "~/components/Swipeable";
 import { ContentPlaceholder } from "~/components/Transition/Placeholder";
 import { SearchResult } from "~/modules/search/components/SearchResult";
 
-/**
- * Sheet allowing us to see the upcoming tracks and remove tracks from
- * the queue.
- */
-export default function TrackUpcomingSheet() {
+/** All the sheets used on `/now-playing` route. */
+export function NowPlayingSheets(
+  props: Record<
+    "playbackOptionsRef" | "upcomingTracksRef",
+    React.RefObject<ActionSheetRef>
+  >,
+) {
+  return (
+    <>
+      <PlaybackOptionsSheet sheetRef={props.playbackOptionsRef} />
+      <TrackUpcomingSheet sheetRef={props.upcomingTracksRef} />
+    </>
+  );
+}
+
+//#region Playback Options
+/** Enables us to specify  how the media is played. */
+function PlaybackOptionsSheet(props: {
+  sheetRef: React.RefObject<ActionSheetRef>;
+}) {
+  const { t } = useTranslation();
+  const playbackSpeed = useSessionStore((state) => state.playbackSpeed);
+  const volume = useSessionStore((state) => state.volume);
+
+  return (
+    <Sheet ref={props.sheetRef} contentContainerClassName="gap-4">
+      <NSlider
+        label={t("feat.playback.extra.speed")}
+        value={playbackSpeed}
+        {...PlaybackSpeedSliderOptions}
+      />
+      <NSlider
+        label={t("feat.playback.extra.volume")}
+        value={volume}
+        {...VolumeSliderOptions}
+      />
+    </Sheet>
+  );
+}
+
+const rateFormatter = new Intl.NumberFormat("en-US", {
+  notation: "compact",
+  maximumFractionDigits: 2,
+});
+
+const PlaybackSpeedSliderOptions = {
+  min: 0.25,
+  max: 2,
+  step: 0.05,
+  trackMarks: [0.25, 0.5, 0.75, 1, 1.25, 1.5, 1.75, 2],
+  icon: <SlowMotionVideo />,
+  onChange: async (playbackSpeed: number) => {
+    sessionStore.setState({ playbackSpeed });
+    await TrackPlayer.setRate(playbackSpeed).catch();
+  },
+  formatValue: (playbackSpeed: number) =>
+    `${rateFormatter.format(playbackSpeed)}x`,
+};
+
+const VolumeSliderOptions = {
+  min: 0,
+  max: 1,
+  step: 0.01,
+  trackMarks: [0, 0.25, 0.5, 0.75, 1],
+  icon: <VolumeUp />,
+  onChange: async (volume: number) => {
+    sessionStore.setState({ volume });
+    await TrackPlayer.setVolume(volume).catch();
+  },
+  formatValue: (volume: number) => `${Math.round(volume * 100)}%`,
+};
+//#endregion
+
+//#region Upcoming Tracks
+/** Enables user to see what tracks are coming up and remove tracks from the queue. */
+function TrackUpcomingSheet(props: {
+  sheetRef: React.RefObject<ActionSheetRef>;
+}) {
   const populateCurrentTrackList = useUpcomingStore(
     (state) => state.populateCurrentTrackList,
   );
@@ -39,15 +117,12 @@ export default function TrackUpcomingSheet() {
       ? trackList.length
       : trackList.length - 1 - listIndex;
 
-  useEffect(() => {
-    populateCurrentTrackList();
-  }, [populateCurrentTrackList]);
-
   return (
     <Sheet
-      id="TrackUpcomingSheet"
+      ref={props.sheetRef}
       titleKey="term.upcoming"
       snapTop
+      onOpen={populateCurrentTrackList}
       contentContainerClassName="px-0"
     >
       <SheetsFlashList
@@ -141,3 +216,4 @@ function TrackItem({
     />
   );
 }
+//#endregion
