+++ mobile/src/screens/NowPlaying/helpers/useVinylSeekbar.ts
@@ -12,7 +12,7 @@ import {
 import TrackPlayer, { useProgress } from "react-native-track-player";
 
 import { useMusicStore } from "~/modules/media/services/Music";
-import { useSeekStore } from "~/screens/NowPlaying/SeekService";
+import { useSeekStore } from "./SeekService";
 
 /** Controls the rotation of the vinyl on the "Now Playing" screen. */
 export function useVinylSeekbar() {
